+++ @remirror/core-utils/src/__tests__/command-utils.spec.ts
@@ -9,15 +9,30 @@ import {
 } from '../command-utils';
 
 describe('removeMark', () => {
-  it('should remove the mark when selected', () => {
+  const type = schema.marks.strong;
+  it('removes the contained mark', () => {
     const from = doc(p(strong('<start>bold<end>')));
     const to = doc(p('bold'));
-    expect(removeMark({ type: schema.marks.strong })).transformsPMNode({ from, to });
+    expect(removeMark({ type })).toTransformNode({ from, to });
   });
 
-  it('should not remove when cursor is within the mark', () => {
+  it('leaves mark untouched when `expand` is `false`', () => {
     const from = doc(p(strong('bo<cursor>ld')));
-    expect(removeMark({ type: schema.marks.strong })).transformsPMNode({ from });
+    expect(removeMark({ type })).toTransformNode({ from });
+  });
+
+  it('removes mark when `expand` is `true`', () => {
+    const from = doc(p(strong('bo<cursor>ld')));
+    const to = doc(p('bold'));
+    expect(removeMark({ type, expand: true })).toTransformNode({ from, to });
+  });
+
+  it('removes the mark from a custom range', () => {
+    const from = doc(p('start ', strong('bold'), ' and not<cursor>'));
+    const to = doc(p('start bold and not'));
+    expect(removeMark({ type, range: { from: 7, to: 11 } })).toTransformNode({ from, to });
+    expect(removeMark({ type, range: { from: 8 }, expand: true })).toTransformNode({ from, to });
+    expect(removeMark({ type, range: { from: 3, to: 7 }, expand: true })).toTransformNode({ from, to });
   });
 });
 
@@ -25,12 +40,12 @@ describe('replaceText', () => {
   it('replaces valid content', () => {
     const from = doc(p('replace <start>me<end>'));
     const to = doc(p('replace ', atomInline()));
-    expect(replaceText({ appendText: '', type: schema.nodes.atomInline })).transformsPMNode({ from, to });
+    expect(replaceText({ appendText: '', type: schema.nodes.atomInline })).toTransformNode({ from, to });
   });
 
   it('does not replace invalid content', () => {
     const from = doc(p('replace <start>me<end>'));
-    expect(replaceText({ appendText: '', type: schema.nodes.heading })).transformsPMNode({ from });
+    expect(replaceText({ appendText: '', type: schema.nodes.heading })).toTransformNode({ from });
   });
 
   it('can specify from and to', () => {
@@ -43,7 +58,7 @@ describe('replaceText', () => {
         content: 'Content',
         range: { from: 8, to: 8 },
       }),
-    ).transformsPMNode({
+    ).toTransformNode({
       from,
       to,
     });
@@ -54,7 +69,7 @@ describe('replaceText', () => {
     const to = doc(p('Ignore'), p('Content '));
     expect(
       replaceText({ appendText: ' ', type: schema.nodes.paragraph, content: 'Content' }),
-    ).transformsPMNode({
+    ).toTransformNode({
       from,
       to,
     });
@@ -64,20 +79,20 @@ describe('replaceText', () => {
 test('updateMark', () => {
   const from = doc(p('Make <start>bold<end>'));
   const to = doc(p('Make ', strong('bold')));
-  expect(updateMark({ type: schema.marks.strong })).transformsPMNode({ from, to });
+  expect(updateMark({ type: schema.marks.strong })).toTransformNode({ from, to });
 });
 
 describe('toggleWrap', () => {
   it('adds the node wrapping the selection', () => {
     const from = doc(p('Wrap ', '<cursor>me'));
     const to = doc(blockquote(p('Wrap me')));
-    expect(toggleWrap(schema.nodes.blockquote)).transformsPMNode({ from, to });
+    expect(toggleWrap(schema.nodes.blockquote)).toTransformNode({ from, to });
   });
 
   it('lifts the node when already wrapped', () => {
     const from = doc(p(blockquote('Lift <cursor>me')));
     const to = doc(blockquote('Lift me'));
-    expect(toggleWrap(schema.nodes.blockquote)).transformsPMNode({ from, to });
+    expect(toggleWrap(schema.nodes.blockquote)).toTransformNode({ from, to });
   });
 });
 
@@ -91,7 +106,7 @@ describe('toggleBlockItem', () => {
         toggleType: schema.nodes.paragraph,
         attrs: { level: 1 },
       }),
-    ).transformsPMNode({
+    ).toTransformNode({
       from,
       to,
     });
@@ -106,7 +121,7 @@ describe('toggleBlockItem', () => {
         toggleType: schema.nodes.paragraph,
         attrs: { level: 1 },
       }),
-    ).transformsPMNode({
+    ).toTransformNode({
       from,
       to,
     });
@@ -117,7 +132,7 @@ describe('toggleList', () => {
   it('toggles to the specified list type', () => {
     const from = doc(p('make <cursor>list'));
     const to = doc(ul(li(p('make list'))));
-    expect(toggleList(schema.nodes.bulletList, schema.nodes.listItem)).transformsPMNode({
+    expect(toggleList(schema.nodes.bulletList, schema.nodes.listItem)).toTransformNode({
       from,
       to,
     });