+++ packages/jest-prosemirror/src/jest-prosemirror-matchers.ts
@@ -5,7 +5,7 @@ import { transformsNodeFailMessage, transformsNodePassMessage } from './jest-pro
 import { CommandTransformation } from './jest-prosemirror-types';
 
 export const prosemirrorMatchers = {
-  transformsPMNode(
+  toTransformNode(
     this: jest.MatcherUtils,
     command: CommandFunction,
     { from, to }: CommandTransformation = Cast({}),
@@ -25,7 +25,7 @@ export const prosemirrorMatchers = {
     const expected = to ? to : from;
     const shouldChange = bool(to);
     const [pass, actual] = apply(from, command, to);
-    const properties = { actual, expected, name: 'transformsPMNode' };
+    const properties = { actual, expected, name: 'toTransformNode' };
 
     if (pass) {
       return { ...properties, pass, message: transformsNodePassMessage(actual, expected, shouldChange) };
@@ -34,11 +34,15 @@ export const prosemirrorMatchers = {
     }
   },
 
-  toEqualPMNode(this: jest.MatcherUtils, actual: TaggedProsemirrorNode, expected: TaggedProsemirrorNode) {
+  toEqualProsemirrorNode(
+    this: jest.MatcherUtils,
+    actual: TaggedProsemirrorNode,
+    expected: TaggedProsemirrorNode,
+  ) {
     const pass = this.equals(actual.toJSON(), expected.toJSON());
     const message = pass
       ? () =>
-          `${this.utils.matcherHint('.not.toEqualPMNode')}\n\n` +
+          `${this.utils.matcherHint('.not.toEqualProsemirrorNode')}\n\n` +
           `Expected JSON value of document to not equal:\n  ${this.utils.printExpected(expected)}\n` +
           `Actual JSON:\n  ${this.utils.printReceived(actual)}`
       : () => {
@@ -46,7 +50,7 @@ export const prosemirrorMatchers = {
             expand: this.expand,
           });
           return (
-            `${this.utils.matcherHint('.toEqualPMNode')}\n\n` +
+            `${this.utils.matcherHint('.toEqualProsemirrorNode')}\n\n` +
             `Expected JSON value of document to equal:\n${this.utils.printExpected(expected)}\n` +
             `Actual JSON:\n  ${this.utils.printReceived(actual)}` +
             `${diffString ? `\n\nDifference:\n\n${diffString}` : ''}`
@@ -58,7 +62,7 @@ export const prosemirrorMatchers = {
       actual,
       expected,
       message,
-      name: 'toEqualPMNode',
+      name: 'toEqualProsemirrorNode',
     };
   },
 };
@@ -67,10 +71,29 @@ declare global {
   namespace jest {
     interface Matchers<R> {
       /**
-       * Test that the correct transformation happens
+       * A utility from jest-prosemirror which tests that a command transforms
+       * the prosemirror node in the desired way.
+       *
+       * ```ts
+       * import { removeMark } from '@remirror/core-utils';
+       * import {schema, doc, p, strong} from 'jest-prosemirror';
+       *
+       * test('remove the mark', () => {
+       *   const type = schema.marks.bold
+       *   const from = doc(p(strong('<start>bold<end>')));
+       *   const to = doc(p('bold'));
+       *
+       *   expect(removeMark({ type })).toTransformNode({ from, to });
+       * });
+       * ```
+       *
+       * This tests that mark has been removed by the provided command.
+       *
+       * The to property is optional and if you would like to test that the node
+       * is identical after the transform you can leave it blank as a shorthand.
        */
-      transformsPMNode(params: CommandTransformation): R;
-      toEqualPMNode(params: ProsemirrorNode): R;
+      toTransformNode(params: CommandTransformation): R;
+      toEqualProsemirrorNode(params: ProsemirrorNode): R;
     }
   }
 }