+++ packages/jest-prosemirror/src/jest-prosemirror-messages.ts
@@ -8,7 +8,7 @@ export const transformsNodePassMessage = (
   expected: TaggedProsemirrorNode,
   shouldChange: boolean,
 ) => () =>
-  matcherHint('.not.transformsPMNode') + '\n\n' + shouldChange
+  matcherHint('.not.toTransformNode') + '\n\n' + shouldChange
     ? chalk`Expected the node {bold not} to be:\n` +
       `${printExpected(expected.toString())}\n` +
       `Position: { from: ${selectionFor(expected).from}, to: ${selectionFor(expected).to} }\n\n` +
@@ -27,7 +27,7 @@ export const transformsNodeFailMessage = (
   expected: TaggedProsemirrorNode,
   shouldChange: boolean,
 ) => () =>
-  matcherHint('.transformsPMNode') + '\n\n' + shouldChange
+  matcherHint('.toTransformNode') + '\n\n' + shouldChange
     ? 'Expected the node to be transformed to:\n' +
       `${printExpected(expected.toString())}\n` +
       `Position: { from: ${selectionFor(expected).from}, to: ${selectionFor(expected).to} }\n\n` +