+++ packages/prosemirror-suggest/src/suggest-constants.ts
@@ -52,9 +52,11 @@ export enum ActionTaken {
  */
 export enum ExitReason {
   /**
-   * The user has pasted some text with multiple characters or run a command that adds multiple characters.
+   * The user has pasted some text with multiple characters or run a command
+   * that adds multiple characters.
    *
-   * `onExit` should be called but the previous match should be retested as it's possible that it's been extended.
+   * `onExit` should be called but the previous match should be retested as it's
+   * possible that it's been extended.
    */
   End = 'exit-end',
 
@@ -64,53 +66,59 @@ export enum ExitReason {
   Removed = 'delete',
 
   /**
-   * The user has pasted some text with multiple characters or run a command that adds multiple characters
-   * somewhere within the active suggestion.
-   * e.g. `@abc` -> `@ab123 asdf aiti c`
+   * The user has pasted some text with multiple characters or run a command
+   * that adds multiple characters somewhere within the active suggestion. e.g.
+   * `@abc` -> `@ab123 asdf aiti c`
    *
-   * `onExit` should be called but the previous match should be retested as it's possible that it's been extended.
+   * `onExit` should be called but the previous match should be retested as it's
+   * possible that it's been extended.
    */
   Split = 'exit-split',
 
   /**
-   * The user has pasted some text with multiple characters or run a command that adds multiple characters
-   * right after the initial multi-character.
-   * e.g. `@abc` -> `@ this is newabc`
+   * The user has pasted some text with multiple characters or run a command
+   * that adds multiple characters right after the initial multi-character. e.g.
+   * `@abc` -> `@ this is newabc`
    *
    * In this case it is best to remove the mention completely.
    */
   InvalidSplit = 'invalid-exit-split',
 
   /**
-   * User has moved out of the suggestion at the end. This can happen via using arrow keys, but can also be
-   * via the suggestion no longer matching as the user types, a mouse click or custom command.
-   * All that has changed is the cursor position.
+   * User has moved out of the suggestion at the end. This can happen via using
+   * arrow keys, but can also be via the suggestion no longer matching as the
+   * user types, a mouse click or custom command. All that has changed is the
+   * cursor position.
    */
   MoveEnd = 'move-end',
 
   /**
-   * User has moved out of the suggestion but from the beginning. This can be via the arrow keys but can also be
-   * via a mouse click or custom command. All that changed is the cursor position.
+   * User has moved out of the suggestion but from the beginning. This can be
+   * via the arrow keys but can also be via a mouse click or custom command. All
+   * that changed is the cursor position.
    */
   MoveStart = 'move-start',
 
   /**
-   * The user has jumped to another suggestion which occurs afterwards in the editor. This can be via a click,
-   * a keyboard jump or custom commands. In this case since there is still an active suggestion it will trigger
-   * both an `onExit` and `onChange` call.
+   * The user has jumped to another suggestion which occurs afterwards in the
+   * editor. This can be via a click, a keyboard jump or custom commands. In
+   * this case since there is still an active suggestion it will trigger both an
+   * `onExit` and `onChange` call.
    */
   JumpForward = 'jump-forward-exit',
 
   /**
-   * The user has jumped to another suggestion which occurs before the previous suggestion in the editor. This can
-   * happen via a click, a keyboard jump (END) or a custom command. In this case since there is still an active
+   * The user has jumped to another suggestion which occurs before the previous
+   * suggestion in the editor. This can happen via a click, a keyboard jump
+   * (END) or a custom command. In this case since there is still an active
    * suggestion it will trigger both an `onExit` and `onChange` call.
    */
   JumpBackward = 'jump-backward-exit',
 
   /**
-   * The user has selected some text outside the current selection, this can trigger
-   * an exit. This can be from a triple click to select the line or Ctrl-A to select all.
+   * The user has selected some text outside the current selection, this can
+   * trigger an exit. This can be from a triple click to select the line or
+   * Ctrl-A to select all.
    */
   SelectionOutside = 'selection-outside',
 }
@@ -125,13 +133,14 @@ export enum ChangeReason {
   Start = 'start',
 
   /**
-   * A changed happened to the character. This can be addition, deletion or replacement.
+   * A changed happened to the character. This can be addition, deletion or
+   * replacement.
    */
   Text = 'change-character',
 
   /**
-   * A change happened to the selection status which was not purely a move.
-   * The selection area may have been increased.
+   * A change happened to the selection status which was not purely a move. The
+   * selection area may have been increased.
    */
   SelectionInside = 'selection-inside',
 
@@ -141,12 +150,14 @@ export enum ChangeReason {
   Move = 'move',
 
   /**
-   * The user has moved from one suggestion to another suggestion earlier in the document.
+   * The user has moved from one suggestion to another suggestion earlier in the
+   * document.
    */
   JumpBackward = 'jump-backward-change',
 
   /**
-   * The user has moved from one suggestion to another suggestion further along in the document.
+   * The user has moved from one suggestion to another suggestion further along
+   * in the document.
    */
   JumpForward = 'jump-forward-change',
 }